Query and list all your submitted tasks

Query and list all your submitted tasks

list_tasks(
  status = c("valid", "running", "init", "finish", "all"),
  order = FALSE,
  expire = 0
)

Arguments

status
filter task status on the server, choices are 'valid', 'init' (submitted, waiting to run), 'running' (running task), 'finish' (finished task), and 'canceled' (canceled by the server)

order
whether to order by date submitted (in descending order); default is false

expire
positive number (in seconds) to filter out tasks that have been submitted most recently, or 0 (default) to list tasks regardless of their dates. For example, expire=100 will only list tasks submitted during the past 100 seconds.

Value

A data frame listing tasks submitted, columns are

name

task name (ID)

userid

your user ID

submitted

1 if submitted and 0 otherwise

collected

1 if result has been collected and 0 otherwise

error

1 if error occurs and 0 otherwise

path

the local directory that stores the task data

serverip

server address if the task has been submitted

serverport

server port if the task has been submitted

removed

1 if the task has been removed and 0 otherwise

time_added

UNIX time of time when task is created. Use as.POSIXct(time_added, origin="1970-01-01") to convert to read-able time; see as.POSIXct