Send requests to the server to query the server status or task status

request_server

sends general request with authentication tokens ready in the request headers

request_task_list

get list of tasks of your current userid from the server

request_server(
  path,
  host = default_host(),
  port = default_port(),
  protocol = default_protocol(),
  body = list(),
  header = list(),
  method = c("POST", "GET"),
  encode = "json"
)

request_task_list(
  task_status = "valid",
  host = default_host(allow0 = FALSE),
  port = default_port(),
  protocol = default_protocol(),
  path = "jobs/list"
)

Arguments

protocol, host, port, path
server location-related configurations, the 'url' of request will be protocol://host:port/path

body
a list of request body

header
additional header key-value pairs

method
method of request; choices are 'POST' and 'GET'

encode
serialization method to encode request body

task_status
task status to filter

Value

'httr' response. You can use content to check the response contents.